requirements.txt
fastapi>=0.104.1
gunicorn>=21.2.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
greenlet>=3.3.0
supabase>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2

# LangGraph and PostgreSQL dependencies
langgraph>=0.2.0
langgraph-checkpoint-postgres>=1.0.0
psycopg[pool]>=3.1.0

